Create a MaskedInputMigrator
